requirements.txt
streamlit==1.50.0
pandas==2.3.2
pillow==11.1.0